converting ivec to vec

+++ ball.cpp
@@ -1,8 +1,8 @@
 #include "ball.h"
 
 Ball::Ball(int x, int y) {
-    position = glm::ivec2(x, y);
-    velocity = glm::ivec2(1, 1);
+    position = glm::vec2(x, y);
+    velocity = glm::vec2(1, 1);
 }
