+++ main.cpp
@@ -16,11 +16,11 @@ int main(int argc, char* argv[]) {
     Window window(640, 480);
     assert(window.init() == 0);
 
-    std::vector<glm::ivec2> vertices = {
-        glm::ivec2(1,1),
-        glm::ivec2(1,9),
-        glm::ivec2(9,9),
-        glm::ivec2(9,1)
+    std::vector<glm::vec2> vertices = {
+        glm::vec2(1,1),
+        glm::vec2(1,9),
+        glm::vec2(9,9),
+        glm::vec2(9,1)
     };
     Course course(10, 10, vertices);
     Ball ball(5, 5);
